Mudando banco de dados para H2

+++ ecommerce-shopping-api/src/main/java/com/rendersonjunior/ecommerceshoppingapi/dto/ShopReportDTO.java
@@ -12,9 +12,8 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ShopReportDTO {
-
     private Long count;
     private BigDecimal total;
-    private BigDecimal mean;
+    private Double mean;
 
 }
